requirements.txt
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu
python-dotenv
numpy
langchain-text-splitters
langchain-core